Get rid of superfluous classes of BaseAttempt
There's really no need for subclasses of the BaseAttempt class.  This
just combines both into a single class. And since there are no longer
multiple implementation of the Attempt interface, the implementation
was moved up into the interface and it was changed to a class.
